onnx_requirements.txt
# ONNX conversion and inference dependencies
onnx>=1.12.0
onnxruntime>=1.12.0
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=8.0.0

# API server dependencies
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0 